+++ src/frontend/Cargo.toml
@@ -25,6 +25,7 @@ clap = { version = "3", features = ["derive"] }
 derivative = "2"
 downcast-rs = "1.2"
 dyn-clone = "1.0.4"
+easy-ext = "1"
 enum-as-inner = "0.5"
 fixedbitset = "0.4.1"
 futures = { version = "0.3", default-features = false, features = ["alloc"] }
