+++ src/frontend/src/optimizer/plan_node/batch_sort_agg.rs
@@ -28,6 +28,7 @@ use super::{
 use crate::expr::{Expr, ExprImpl, ExprRewriter, InputRef};
 use crate::optimizer::plan_node::ToLocalBatch;
 use crate::optimizer::property::{Distribution, Order, RequiredDist};
+use crate::utils::ColIndexMappingRewriteExt;
 
 #[derive(Debug, Clone)]
 pub struct BatchSortAgg {
